% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biclustering.R
\name{generate_example_matrix}
\alias{generate_example_matrix}
\title{Generate an example matrix}
\usage{
generate_example_matrix()
}
\value{
A list with length 2:
original:  A numeric 14 x 14 matrix with a checkerboard structure.
randomized:  The randomized matrix.
}
\description{
Generates a numeric matrix, simulating correlation coefficients in a dataset, and randomizes it.
}
